requirements.txt
numpy
marshmallow_dataclass
shapely